Add files via upload
The objective of this project was to create a mobile application to assist users in managing their food inventory. The application would store and display basic information about the inventory contents and alerts the user of the food products which are due to expire the next day. Consequentially, users may take actions to avoid the concerned products get wasted or spoiled. It is believed that a considerable amount of food waste would be avoided in households if the occupants are well-informed of the timeline of their food stocks. Provisions have also been made to allow for the multi-device use. 
Most food management applications available are mainly concerned with helping users watch their weight and food in-take and generally requires lots of information from user. The advantage of this project is the use of the simplest information of food products to monitor the inventory. With an eye on the future, a demo solution was integrated to show compatibility with future advancement in food packaging.